Add fallback BinCurlHttpClient
